backend/requirements.txt
djangorestframework
Django
